update the sample rate to be a floating point number

+++ app/models.py
@@ -8,7 +8,7 @@ from sqlmodel import Field, Relationship, SQLModel
 class Session(SQLModel, table=True):
     id: uuid.UUID = Field(default_factory=uuid.uuid4, primary_key=True)
     user_agent: str = Field(max_length=1024)
-    sample_rate: int = Field()
+    sample_rate: float = Field()
     created_at: datetime.datetime = Field(default_factory=lambda: datetime.datetime.now(datetime.timezone.utc))
     duration: float = Field()
     samples: list["Sample"] = Relationship(back_populates="session")
@@ -27,7 +27,7 @@ class Sample(SampleBase, table=True):
 
 # request/response schemas
 class SessionCreate(BaseModel):
-    sample_rate: int
+    sample_rate: float
     recorded_at: datetime.datetime | None = None
     duration: float
     samples: List[SampleBase]
@@ -45,7 +45,7 @@ class SessionRead(BaseModel):
     id: uuid.UUID
     user_agent: str
     duration: float
-    sample_rate: int
+    sample_rate: float
     created_at: datetime.datetime
